refactor(loader): unify asset schemas under schema directory
