+++ packages/loader/src/PrefabLoader.ts
@@ -1,7 +1,7 @@
 import { AssetPromise, AssetType, Loader, LoadItem, resourceLoader, ResourceManager } from "@galacean/engine-core";
 import { PrefabParser } from "./prefab/PrefabParser";
 import { PrefabResource } from "./prefab/PrefabResource";
-import type { PrefabFile } from "./scene-format/types";
+import type { PrefabFile } from "./schema/PrefabSchema";
 
 @resourceLoader(AssetType.Prefab, ["prefab"])
 export class PrefabLoader extends Loader<PrefabResource> {
