+++ packages/loader/src/index.ts
@@ -33,6 +33,5 @@ export * from "./gltf";
 export { KTX2Loader } from "./ktx2/KTX2Loader";
 export { KTX2TargetFormat } from "./ktx2/KTX2TargetFormat";
 export * from "./resource-deserialize";
-export * from "./scene-format/types";
 export * from "./prefab/PrefabResource";
 export { HDRDecoder } from "./HDRDecoder";
