--- packages/loader/src/resource-deserialize/resources/schema/index.ts
@@ -1,3 +1,0 @@
-export * from "./BasicSchema";
-export * from "./MaterialSchema";
-export * from "./ProjectSchema";
