+++ packages/loader/src/schema/PrefabSchema.ts
@@ -1,0 +1,5 @@
+import type { HierarchyFile } from "./HierarchySchema";
+
+export interface PrefabFile extends HierarchyFile {
+  root: number;
+}
