+++ packages/loader/src/schema/index.ts
@@ -1,0 +1,7 @@
+export * from "./CommonSchema";
+export * from "./HierarchySchema";
+export * from "./SceneSchema";
+export * from "./PrefabSchema";
+export * from "./BasicSchema";
+export * from "./MaterialSchema";
+export * from "./ProjectSchema";
